misc: doc web polyfill
