+++ lib/java/ydoc-server/src/main/java/org/enso/ydoc/polyfill/web/AbortController.java
@@ -1,4 +1,4 @@
-package org.enso.ydoc.polyfill.nodejs;
+package org.enso.ydoc.polyfill.web;
 
 import java.util.Arrays;
 import java.util.UUID;
@@ -8,6 +8,11 @@ import org.graalvm.polyglot.Source;
 import org.graalvm.polyglot.Value;
 import org.graalvm.polyglot.proxy.ProxyExecutable;
 
+/**
+ * Implements the <a
+ * href="https://nodejs.org/api/globals.html#class-abortcontroller">AbortController</a> Node.js
+ * interface.
+ */
 final class AbortController implements ProxyExecutable, Polyfill {
 
   private static final String RANDOM_UUID = "random-uuid";
