+++ lib/java/ydoc-server/src/main/resources/org/enso/ydoc/polyfill/web/text-decoder.js
