+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/polyfill/web/AbortControllerTest.java
@@ -1,30 +1,27 @@
-package org.enso.ydoc.polyfill.nodejs;
+package org.enso.ydoc.polyfill.web;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import org.graalvm.polyglot.Context;
-import org.graalvm.polyglot.HostAccess;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-public class EventEmitterTest {
+public class AbortControllerTest {
 
   private Context context;
   private ExecutorService executor;
 
-  public EventEmitterTest() {}
+  public AbortControllerTest() {}
 
   @Before
   public void setup() throws Exception {
     executor = Executors.newSingleThreadExecutor();
     var eventTarget = new EventTarget(executor);
-    var eventEmitter = new EventEmitter();
-
-    var hostAccess = HostAccess.newBuilder(HostAccess.EXPLICIT).allowArrayAccess(true).build();
-    var b = Context.newBuilder("js").allowHostAccess(hostAccess);
+    var abortController = new AbortController();
+    var b = Context.newBuilder("js");
 
     var chromePort = Integer.getInteger("inspectPort", -1);
     if (chromePort > 0) {
@@ -36,7 +33,7 @@ public class EventEmitterTest {
                 () -> {
                   var ctx = b.build();
                   eventTarget.initialize(ctx);
-                  eventEmitter.initialize(ctx);
+                  abortController.initialize(ctx);
                   return ctx;
                 },
                 executor)
@@ -50,35 +47,30 @@ public class EventEmitterTest {
   }
 
   @Test
-  public void emit() throws Exception {
+  public void getSignal() throws Exception {
     var code =
         """
-        var count = 0;
-        var ee = new EventEmitter();
-        ee.on('inc', (a, b) => count += 10*a + b);
-        ee.emit('inc', 4, 2);
-        count;
+        var ctrl = new AbortController();
+        ctrl.signal.aborted;
         """;
 
     var result = CompletableFuture.supplyAsync(() -> context.eval("js", code), executor).get();
 
-    Assert.assertEquals(42, result.asInt());
+    Assert.assertFalse(result.asBoolean());
   }
 
   @Test
-  public void listeners() throws Exception {
+  public void abort() throws Exception {
     var code =
         """
-        var count = 0;
-        var ee = new EventEmitter();
-        ee.on('inc', (a, b) => count += 10*a + b);
-        ee.listeners('inc')
+        var ctrl = new AbortController();
+        var signal = ctrl.signal;
+        ctrl.abort();
+        signal.aborted;
         """;
 
     var result = CompletableFuture.supplyAsync(() -> context.eval("js", code), executor).get();
-    var arr = result.as(Object[].class);
 
-    Assert.assertEquals(1, arr.length);
-    Assert.assertNotNull(arr[0]);
+    Assert.assertTrue(result.asBoolean());
   }
 }
