+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/polyfill/web/EventEmitterTest.java
@@ -1,27 +1,30 @@
-package org.enso.ydoc.polyfill.nodejs;
+package org.enso.ydoc.polyfill.web;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import org.graalvm.polyglot.Context;
+import org.graalvm.polyglot.HostAccess;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-public class AbortControllerTest {
+public class EventEmitterTest {
 
   private Context context;
   private ExecutorService executor;
 
-  public AbortControllerTest() {}
+  public EventEmitterTest() {}
 
   @Before
   public void setup() throws Exception {
     executor = Executors.newSingleThreadExecutor();
     var eventTarget = new EventTarget(executor);
-    var abortController = new AbortController();
-    var b = Context.newBuilder("js");
+    var eventEmitter = new EventEmitter();
+
+    var hostAccess = HostAccess.newBuilder(HostAccess.EXPLICIT).allowArrayAccess(true).build();
+    var b = Context.newBuilder("js").allowHostAccess(hostAccess);
 
     var chromePort = Integer.getInteger("inspectPort", -1);
     if (chromePort > 0) {
@@ -33,7 +36,7 @@ public class AbortControllerTest {
                 () -> {
                   var ctx = b.build();
                   eventTarget.initialize(ctx);
-                  abortController.initialize(ctx);
+                  eventEmitter.initialize(ctx);
                   return ctx;
                 },
                 executor)
@@ -47,30 +50,35 @@ public class AbortControllerTest {
   }
 
   @Test
-  public void getSignal() throws Exception {
+  public void emit() throws Exception {
     var code =
         """
-        var ctrl = new AbortController();
-        ctrl.signal.aborted;
+        var count = 0;
+        var ee = new EventEmitter();
+        ee.on('inc', (a, b) => count += 10*a + b);
+        ee.emit('inc', 4, 2);
+        count;
         """;
 
     var result = CompletableFuture.supplyAsync(() -> context.eval("js", code), executor).get();
 
-    Assert.assertFalse(result.asBoolean());
+    Assert.assertEquals(42, result.asInt());
   }
 
   @Test
-  public void abort() throws Exception {
+  public void listeners() throws Exception {
     var code =
         """
-        var ctrl = new AbortController();
-        var signal = ctrl.signal;
-        ctrl.abort();
-        signal.aborted;
+        var count = 0;
+        var ee = new EventEmitter();
+        ee.on('inc', (a, b) => count += 10*a + b);
+        ee.listeners('inc')
         """;
 
     var result = CompletableFuture.supplyAsync(() -> context.eval("js", code), executor).get();
+    var arr = result.as(Object[].class);
 
-    Assert.assertTrue(result.asBoolean());
+    Assert.assertEquals(1, arr.length);
+    Assert.assertNotNull(arr[0]);
   }
 }
