+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/polyfill/web/WebSocketTest.java
@@ -1,4 +1,4 @@
-package org.enso.ydoc.polyfill.nodejs;
+package org.enso.ydoc.polyfill.web;
 
 import io.helidon.common.buffers.BufferData;
 import io.helidon.webserver.WebServer;
@@ -61,7 +61,7 @@ public class WebSocketTest {
         CompletableFuture.supplyAsync(
                 () -> {
                   var ctx = b.build();
-                  NodeJs.initialize(ctx, executor);
+                  WebEnvironment.initialize(ctx, executor);
                   return ctx;
                 },
                 executor)
